# src/main/resources/openapi/openapi-spec.yaml
openapi: 3.0.3
info:
  title: Wallet Service API
  description: |
    # Digital Wallet Service API
    
    A comprehensive REST API for managing digital wallets, transactions, and users.
    
    ## Features
    
    - **Wallet Management**: Create, retrieve, and manage digital wallets
    - **Transactions**: Deposit, withdraw, and transfer funds
    - **User Management**: User registration and authentication
    - **Transaction History**: Query and filter transaction records
    - **Notifications**: Real-time notifications via Kafka
    
    ## Authentication
    
    All API endpoints (except auth endpoints) require JWT Bearer token authentication.
    Include the token in the Authorization header:
    
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limiting
    
    - 100 requests per second per IP
    - 1000 requests per minute per user
    
    ## Error Codes
    
    | Code | Description |
    |------|-------------|
    | 400  | Bad Request - Invalid input |
    | 401  | Unauthorized - Invalid or missing token |
    | 403  | Forbidden - Insufficient permissions |
    | 404  | Not Found - Resource not found |
    | 409  | Conflict - Resource already exists |
    | 422  | Unprocessable Entity - Validation failed |
    | 500  | Internal Server Error |
    
  version: 1.0.0
  contact:
    name: Wallet Service Team
    email: support@wallet.service
    url: https://wallet.service/support
  license:
    name: Apache 2.0
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /transactions/{transactionId}:
    get:
      tags:
        - Transactions
      summary: Get transaction by ID
      description: Retrieves a specific transaction
      operationId: getTransaction
      parameters:
        - $ref: '#/components/parameters/TransactionIdParam'
      responses:
        '200':
          description: Transaction details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '404':
          description: Transaction not found
  /transactions/wallet/{walletId}:
    get:
      tags:
        - Transactions
      summary: Get transactions by wallet
      description: All transactions where wallet is sender or receiver
      operationId: getTransactionsByWallet
      parameters:
        - $ref: '#/components/parameters/WalletIdParam'
      responses:
        '200':
          description: Transactions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransactionDTO'
  /transactions/wallet/{walletId}/status/{status}:
    get:
      tags:
        - Transactions
      summary: Get transactions by wallet and status
      operationId: getTransactionsByWalletAndStatus
      parameters:
        - $ref: '#/components/parameters/WalletIdParam'
        - $ref: '#/components/parameters/StatusParam'
      responses:
        '200':
          description: Filtered transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransactionDTO'
  /transactions/date-range:
    get:
      tags:
        - Transactions
      summary: Get transactions by date range
      operationId: getTransactionsByDateRange
      parameters:
        - $ref: '#/components/parameters/StartDateParam'
        - $ref: '#/components/parameters/EndDateParam'
      responses:
        '200':
          description: Transactions in date range
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransactionDTO'
  /transactions/amount-range:
    get:
      tags:
        - Transactions
      summary: Get transactions by amount range
      operationId: getTransactionsByAmountRange
      parameters:
        - name: minAmount
          in: query
          required: true
          description: Minimum amount
          schema:
            type: number
            format: decimal
            example: 10.00
        - name: maxAmount
          in: query
          required: true
          description: Maximum amount
          schema:
            type: number
            format: decimal
            example: 1000.00
      responses:
        '200':
          description: Filtered transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransactionDTO'
  /transactions/stats/wallet/{walletId}:
    get:
      tags:
        - Transactions
      summary: Get wallet transaction statistics
      operationId: getWalletTransactionStats
      parameters:
        - $ref: '#/components/parameters/WalletIdParam'
      responses:
        '200':
          description: Transaction statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionStats'
  /transactions/recent:
    get:
      tags:
        - Transactions
      summary: Get recent transactions
      operationId: getRecentTransactions
      parameters:
        - name: limit
          in: query
          description: Number of transactions
          schema:
            type: integer
            default: 10
            example: 20
      responses:
        '200':
          description: Recent transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransactionDTO'
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint

  # =============================================================================
  # Parameters
  # =============================================================================
  parameters:
    WalletIdParam:
      name: walletId
      in: path
      required: true
      description: Unique identifier of the wallet
      schema:
        type: string
        format: uuid
        example: 550e8400-e29b-41d4-a716-446655440000
...
    TransactionIdParam:
      name: transactionId
      in: path
      required: true
      description: Unique identifier of the transaction
      schema:
        type: string
        format: uuid
        example: 15c5713b-b915-4ba3-a101-a7ee931898ac
    StatusParam:
      name: status
      in: path
      required: true
      description: Transaction status
      schema:
        type: string
        enum: [PENDING, COMPLETED, FAILED, CANCELLED]
        example: COMPLETED
    StartDateParam:
      name: startDate
      in: query
      required: true
      description: Start date (ISO-8601)
      schema:
        type: string
        format: date-time
        example: 2024-01-01T00:00:00Z
    EndDateParam:
      name: endDate
      in: query
      required: true
      description: End date (ISO-8601)
      schema:
        type: string
        format: date-time
        example: 2024-12-31T23:59:59Z
...
  schemas:
...
    TransactionResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Deposit completed successfully
        data:
          $ref: '#/components/schemas/TransactionDTO'
        statusCode:
          type: integer
          example: 200
    TransactionDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 15c5713b-b915-4ba3-a101-a7ee931898ac
        fromWalletId:
          type: string
          format: uuid
          nullable: true
          example: 550e8400-e29b-41d4-a716-446655440000
        toWalletId:
          type: string
          format: uuid
          nullable: true
          example: 550e8400-e29b-41d4-a716-446655440001
        amount:
          type: number
          format: decimal
          example: 100.00
        status:
          type: string
          enum: [PENDING, COMPLETED, FAILED, CANCELLED]
          example: COMPLETED
        createdAt:
          type: string
          format: date-time
          example: 2024-01-01T12:00:00Z
    TransactionStats:
      type: object
      properties:
        walletId:
          type: string
          format: uuid
        totalAmount:
          type: number
          format: decimal
        totalSent:
          type: number
          format: decimal
        totalReceived:
          type: number
          format: decimal
        transactionCount:
          type: integer
          format: int64

    # Error Schemas
...